# openapi.yaml
openapi: 3.1.0
info:
  title: 山海灵感便利店领域 API
  version: 2.0.0
  description: App 用户接口与管理端事务动作。字段统一 snake_case，金额统一十进制字符串。
servers:
  - url: http://localhost:3000/api/v1
...
paths:
...
  /orders:
    post:
      tags: [Orders]
      security: [{ UserBearer: [] }]
      summary: 事务创建订单
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody: { required: true, content: { application/json: { schema: { allOf: [{ $ref: "#/components/schemas/CheckoutInput" }, { type: object, required: [address_id], properties: { address_id: { $ref: "#/components/schemas/Id" }, remark: { type: string, maxLength: 255 } } }] } } } }
      responses: { "201": { $ref: "#/components/responses/Success" }, "409": { $ref: "#/components/responses/Error" } }
    get:
      tags: [Orders]
      security: [{ UserBearer: [] }]
      summary: 我的订单分页
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [all, pending-payment, pending-shipment, pending-receipt, completed, cancelled, after-sale] } }
        - { name: page, in: query, schema: { type: integer } }
        - { name: pageSize, in: query, schema: { type: integer, maximum: 100 } }
      responses: { "200": { $ref: "#/components/responses/List" } }
  /orders/counts:
    get:
      tags: [Orders]
      security: [{ UserBearer: [] }]
      summary: 我的订单状态数量
      responses: { "200": { $ref: "#/components/responses/Success" } }
  /orders/{orderNo}:
    get:
      tags: [Orders]
      security: [{ UserBearer: [] }]
      parameters: [{ $ref: "#/components/parameters/OrderNo" }]
      summary: 本人订单完整详情
      responses: { "200": { $ref: "#/components/responses/Success" }, "404": { $ref: "#/components/responses/Error" } }
  /orders/{orderNo}/cancel:
    post:
      tags: [Orders]
      security: [{ UserBearer: [] }]
      parameters: [{ $ref: "#/components/parameters/OrderNo" }]
      summary: 取消待付款订单并恢复库存和券
      requestBody: { content: { application/json: { schema: { type: object, properties: { reason: { type: string } } } } } }
      responses: { "200": { $ref: "#/components/responses/Success" }, "409": { $ref: "#/components/responses/Error" } }
  /orders/{orderNo}/confirm-receipt:
    post:
      tags: [Orders]
      security: [{ UserBearer: [] }]
      parameters: [{ $ref: "#/components/parameters/OrderNo" }]
      summary: 确认收货
      responses: { "200": { $ref: "#/components/responses/Success" }, "409": { $ref: "#/components/responses/Error" } }
...
components:
  securitySchemes:
    UserBearer: { type: http, scheme: bearer, description: 登录返回的 App 会话 Token }
    AdminBearer: { type: http, scheme: bearer, description: 环境变量 API_TOKEN }
  parameters:
...
    OrderNo: { name: orderNo, in: path, required: true, schema: { type: string } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, maxLength: 128 } }
  schemas:
    Id: { oneOf: [{ type: string, pattern: '^\d+$' }, { type: integer, minimum: 1 }] }
...
    CheckoutInput:
      type: object
      description: cart_item_ids 与 item 必须二选一
      properties:
        cart_item_ids: { type: array, minItems: 1, items: { $ref: "#/components/schemas/Id" } }
        item:
          type: object
          required: [product_id, qty]
          properties:
            product_id: { $ref: "#/components/schemas/Id" }
            sku_id: { oneOf: [{ $ref: "#/components/schemas/Id" }, { type: "null" }] }
            qty: { type: integer, minimum: 1 }
        address_id: { $ref: "#/components/schemas/Id" }
        user_coupon_id: { $ref: "#/components/schemas/Id" }
    Error:
      type: object
      properties:
        error:
          type: object
          properties: { code: { type: string }, message: { type: string }, details: {} }
        requestId: { type: string }
  responses:
    Success:
      description: 成功
      content: { application/json: { schema: { type: object, properties: { data: {} } } } }
    List:
      description: 分页成功
      content:
        application/json:
          schema:
            type: object
            properties:
              data: { type: array, items: {} }
              meta: { type: object, properties: { page: { type: integer }, pageSize: { type: integer }, total: { type: integer } } }
    Error:
      description: 请求失败
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }